fix: estimation box component
