+++ src/routes/(console)/create-organization/+page.svelte
@@ -252,7 +252,6 @@
             <EstimatedTotalBox
                 billingPlan={selectedPlan}
                 {collaborators}
-                plans={data.plansInfo}
                 bind:couponData={selectedCoupon}
                 bind:billingBudget />
         {:else}
